+++ NEWS.md
@@ -1,5 +1,7 @@
 # openairmaps (development version)
 
+* `networkMap()` now uses different coloured markers for different networks. If more than one network is specified, and `draw.legend` is set to `TRUE`, a legend will also be drawn for quick identification of different data sources.
+
 * Fixed issues where multiple `addPolarMarkers()` chained together would all show the same plot.
 
 * Fixed issue where `...` and `pollutant` weren't being passed to `addPolarMarkers()`.